feat: transformer + readme instructions
