+++ .gitignore
@@ -165,6 +165,7 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
-# Trained models weights
-trained_model/*
-logs/*
+# Transformer trained model
+transformers/trained_model/
+# event tf
+events.out.tfevents.*
